#4- api 명세서에 맞게 uri 변경
- write -> create로 통일
- 확인을 위한 요청은 __-or-not 형식으로 통일
- 기본적인 리스트 조회는 /로 통일 
- etc.
